Made changes to login_system.cs
Added ways to verify if the user tries to register a new user, checks if
there is a user with that name already. Also checks so that the user
input isn't empty. And when a user logs in checks if the username and
password matches someone in the list. Also made it possible for the
program to save information about the new users and save them in a
usertextfile between runs.