Added code for the user to book an appointment
By taking in the date and time, creating an event of the booking and
confirming it to the user. Next step is for the user to see their
upcoming appointments and for after an appointment for it to show up in
the journal. Also tested out logic to subscribe and unsubscribe to the
event. Will add more code to it in the future for example when its time
to connect the appointment to a dr/nurses schedule, or for the journal.