Changed from events to eventdriven logic
Once again mixed up what was asked in the assignment and after
discussions with counceler about event and the other counceler talking
about event driven logic. At first i made the program on events, however
we only needed the program to have the eventdriven logic. Changed the
code from before, after that i wrote the logic for the user to be able
to see their upcoming appointments.