Added logic to save to journal.txt
Made a savejournal void in our save_system, which writes in the patient
name, dr name, date of the note being made and the actual note int the
system. Next step is to be able to read the data from the textfiles and
use it when a user wants to see their journal.