Added pending role, completed case 6
I have now added a role Pending which users automatically gets when
creating an account. This user gains the role Patient after an admin
have chosen to accept them in "handle registrations". After adding the
pending role I have completed the accept/deny function in handle
registration and it now changes the users role from Pending to Patient
in the users.txt if accepted. If denied that user is deleted from
users.txt.